requirements.txt
django
djangorestframework
psycopg2-binary
Pillow
django-filter
python-decouple
djangorestframework_simplejwt
python-slugify
wheel
drf-yasg
cloudinary